# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutResponseDto": {
        "properties": {
          "email": { "format": "email", "type": "string" },
          "id": { "format": "uuid", "type": "string" },
          "status": {
            "enum": ["OPEN", "AWAITING_PAYMENT", "COMPLETED", "EXPIRED", "CANCELLED"],
            "type": "string"
          }
        },
        "required": ["id", "status", "email"],
        "type": "object"
      },
...
      "ConfirmCheckoutResponseDto": {
        "properties": {
          "checkout": { "$ref": "#/components/schemas/CheckoutResponseDto" },
          "orderNumber": { "example": "HNY-2026-000123", "type": "string" }
        },
        "required": ["checkout", "orderNumber"],
        "type": "object"
      },
...
      "CustomerOrderResponseDto": {
        "properties": {
          "grandTotal": { "$ref": "#/components/schemas/MoneyDto" },
          "number": { "example": "HNY-2026-000123", "type": "string" },
          "status": { "type": "string" }
        },
        "required": ["number", "status", "grandTotal"],
        "type": "object"
      },
...
      "MoneyDto": {
        "properties": {
          "amountMinor": { "example": "125000", "type": "string" },
          "currency": { "example": "IRR", "type": "string" }
        },
        "required": ["amountMinor", "currency"],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "detail": { "example": "The request could not be processed.", "type": "string" },
          "errors": {
            "items": { "$ref": "#/components/schemas/ValidationIssueDto" },
            "type": "array"
          },
          "instance": { "example": "/v1/example", "type": "string" },
          "requestId": { "example": "018f5d36-7b89-7a67-bb7a-e8f9f5db7412", "type": "string" },
          "status": { "example": 422, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://api.honey.invalid/problems/validation-failed",
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "code", "instance", "requestId"],
        "type": "object"
      },
...
      "ValidationIssueDto": {
        "properties": {
          "code": { "example": "INVALID_VALUE", "type": "string" },
          "path": { "example": "field", "type": "string" }
        },
        "required": ["path", "code"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "description": "Opaque server-side session. Local HTTP development uses honey_session.",
        "in": "cookie",
        "name": "__Host-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Honey API contract. Cookie-authenticated unsafe requests require the X-CSRF-Token double-submit header.",
    "title": "Honey API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/checkout": {
      "post": {
        "description": "Creates a checkout from the server-derived cart owner. The client may send contact and address selections only; money and owner fields are rejected.",
        "operationId": "startCheckout",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CheckoutResponseDto" }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Start an owner-scoped checkout",
        "tags": ["Checkout"]
      }
    },
    "/v1/checkout/{id}": {
      "get": {
        "description": "Returns the caller-owned checkout. Another owner receives 404.",
        "operationId": "getCheckout",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CheckoutResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Read the current owner checkout",
        "tags": ["Checkout"]
      }
    },
    "/v1/checkout/{id}/confirm": {
      "post": {
        "description": "Creates exactly one UNPAID order from the locked checkout snapshot. Replays with the same Idempotency-Key return the same order.",
        "operationId": "confirmCheckout",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ConfirmCheckoutResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Create one pending-payment order",
        "tags": ["Checkout"]
      }
    },
    "/v1/checkout/{id}/extend": {
      "post": {
        "description": "The server computes the new expiry. The client sends no body and cannot set the hold length.",
        "operationId": "extendCheckoutReservation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CheckoutResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Extend an active checkout reservation hold once, up to the 30-minute maximum",
        "tags": ["Checkout"]
      }
    },
...
    "/v1/orders/{number}": {
      "get": {
        "description": "Returns one owner-scoped order. Another owner or a guessed order number receives 404.",
        "operationId": "getMyOrder",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": { "example": "HNY-2026-000123", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerOrderResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "summary": "Read one owner-scoped order",
        "tags": ["Orders"]
      }
    },
...
  },
  "servers": [],
...
}